requirements.txt
cryptography==42.0.8
argon2-cffi
Faker
Flask
gevent
psutil
pycryptodome
pyperclip
Requests
requests_toolbelt
dnspython
PyQt6
PyQt6-WebEngine
loguru
aiohttp
aiofiles
xxhash
zmq
protobuf>=5,<7